third_party/blink/web_tests/platform/mac/external/wpt/css/css-fonts/cjk-kerning-expected.txt

This is a testharness.js-based test.
[FAIL] expected match: .kernON .cjk vs .paltONkernON .cjk
  assert_equals: expected 330 but got 378
[FAIL] expected match: .default .cjk vs .kernOFF .cjk
  assert_equals: expected 384 but got 378
Harness: the test ran to completion.
